pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "openai>=1.60.0",
    "anthropic>=0.43.0",
    "google-genai>=1.5.0",
    "asyncpg>=0.29.0",
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-multipart>=0.0.7",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
]
